+++ src/Form/LotteryType.php
@@ -15,6 +15,7 @@ use Symfony\Component\Form\AbstractType;
 use c975L\ShopBundle\Form\LotteryPrizeType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\DateTimeType;
 use Symfony\Component\Form\Extension\Core\Type\CollectionType;
@@ -27,6 +28,13 @@ class LotteryType extends AbstractType
             ->add('isActive', CheckboxType::class, [
                 'label' => 'label.enable_lottery',
                 'required' => false
+            ])
+            ->add('identifier', TextType::class, [
+                'label' => 'label.lottery_identifier',
+                'required' => false,
+                'attr' => [
+                    'readonly' => true
+                ],
             ])
             ->add('drawDate', DateTimeType::class, [
                 'label' => 'label.draw_date',